Rename policy::PortPolicies to policy::Store
Move the store into its own module to help hide/isolate implementation
details.

Move discover/initialization into Store::discover.
