+++ linkerd/app/inbound/src/policy/config.rs
@@ -1,16 +1,8 @@
-use super::{
-    discover::{self, Discover},
-    DefaultPolicy, PortMap, PortPolicies, ServerPolicy,
-};
-use futures::prelude::*;
+use super::{discover::Discover, DefaultPolicy, ServerPolicy, Store};
 use linkerd_app_core::{
-    control, dns, metrics,
-    proxy::{http, identity::LocalCrtKey},
-    svc::NewService,
-    Error, Result,
+    control, dns, metrics, proxy::identity::LocalCrtKey, svc::NewService, Result,
 };
-use std::{collections::HashSet, sync::Arc};
-use tokio::sync::watch;
+use std::collections::{HashMap, HashSet};
 
 /// Configures inbound policies.
 ///
@@ -26,7 +18,7 @@ pub enum Config {
     },
     Fixed {
         default: DefaultPolicy,
-        ports: PortMap<ServerPolicy>,
+        ports: HashMap<u16, ServerPolicy>,
     },
 }
 
@@ -38,24 +30,9 @@ impl Config {
         dns: dns::Resolver,
         metrics: metrics::ControlHttp,
         identity: Option<LocalCrtKey>,
-    ) -> Result<PortPolicies> {
+    ) -> Result<Store> {
         match self {
-            Self::Fixed { default, ports } => {
-                let rxs = ports
-                    .into_iter()
-                    .map(|(p, s)| {
-                        // When using a fixed policy, we don't need to watch for changes. It's
-                        // safe to discard the sender, as the receiver will continue to let us
-                        // borrow/clone each fixed policy.
-                        let (_, rx) = watch::channel(Arc::new(s));
-                        (p, rx)
-                    })
-                    .collect();
-                Ok(PortPolicies {
-                    default,
-                    ports: Arc::new(rxs),
-                })
-            }
+            Self::Fixed { default, ports } => Ok(Store::fixed(default, ports)),
             Self::Discover {
                 control,
                 ports,
@@ -67,39 +44,8 @@ impl Config {
                     let c = control.build(dns, metrics, identity).new_service(());
                     Discover::new(workload, c).into_watch(backoff)
                 };
-                let rxs = Self::spawn_watches(discover, ports).await?;
-                Ok(PortPolicies {
-                    default,
-                    ports: Arc::new(rxs),
-                })
+                Store::discover(default, discover, ports).await
             }
-        }
-    }
-
-    // XXX(ver): rustc can't seem to figure out that this Future is `Send` unless we annotate it
-    // explicitly, hence the manual_async_fn.
-    #[allow(clippy::manual_async_fn)]
-    fn spawn_watches<S>(
-        discover: discover::Watch<S>,
-        ports: HashSet<u16>,
-    ) -> impl Future<Output = Result<PortMap<watch::Receiver<Arc<ServerPolicy>>>, tonic::Status>> + Send
-    where
-        S: tonic::client::GrpcService<tonic::body::BoxBody, Error = Error>,
-        S: Clone + Send + Sync + 'static,
-        S::Future: Send,
-        S::ResponseBody: http::HttpBody<Error = Error> + Send + Sync + 'static,
-    {
-        async move {
-            let rxs = ports.into_iter().map(|port| {
-                discover
-                    .clone()
-                    .spawn_watch(port)
-                    .map_ok(move |rsp| (port, rsp.into_inner()))
-            });
-            futures::future::join_all(rxs)
-                .await
-                .into_iter()
-                .collect::<Result<PortMap<_>, tonic::Status>>()
         }
     }
 }
